requirements.txt
Flask==3.0.3
requests==2.32.5
python-dotenv==1.0.1
gunicorn==22.0.0
